--- .eslintrc.json
@@ -1,80 +1,0 @@
-{
-	"parser": "@typescript-eslint/parser",
-	"parserOptions": {
-		"ecmaVersion": 2022,
-		"sourceType": "module",
-		"project": "./tsconfig.lint.json"
-	},
-	"ignorePatterns": ["node_modules"],
-	"plugins": ["@typescript-eslint", "vitest", "import"],
-	"extends": [
-		"eslint:recommended",
-		"plugin:@typescript-eslint/recommended",
-		"plugin:@typescript-eslint/recommended-requiring-type-checking",
-		"plugin:import/recommended",
-		"plugin:import/typescript",
-		"plugin:vitest/recommended",
-		"plugin:import-esm/recommended"
-	],
-	"rules": {
-		"@typescript-eslint/no-empty-interface": "warn",
-		"@typescript-eslint/ban-ts-comment": "off",
-		"@typescript-eslint/no-use-before-define": "off",
-		"@typescript-eslint/no-non-null-assertion": "warn",
-		"@typescript-eslint/no-var-requires": "off",
-		"@typescript-eslint/indent": "off",
-		"@typescript-eslint/no-explicit-any": "warn",
-		"@typescript-eslint/no-unsafe-member-access": "warn",
-		"@typescript-eslint/explicit-function-return-type": "off",
-		"@typescript-eslint/consistent-type-imports": "warn",
-		"@typescript-eslint/no-unused-vars": [
-			"warn",
-			{
-				"argsIgnorePattern": "^_",
-				"varsIgnorePattern": "^_",
-				"caughtErrorsIgnorePattern": "^_"
-			}
-		],
-		"@typescript-eslint/member-delimiter-style": [
-			"error",
-			{
-				"multiline": {
-					"delimiter": "none",
-					"requireLast": false
-				},
-				"singleline": {
-					"delimiter": "comma",
-					"requireLast": false
-				}
-			}
-		],
-		"import/no-default-export": "error",
-		"import/order": [
-			"warn",
-			{
-				"alphabetize": { "order": "asc" },
-				"newlines-between": "always"
-			}
-		],
-		"import/no-unresolved": ["off"],
-		"object-shorthand": ["error", "always"],
-		"max-lines": ["error", { "max": 600 }],
-		"max-params": ["error", { "max": 4 }],
-		"max-statements": ["error", { "max": 15 }],
-		"complexity": ["error", { "max": 20 }]
-	},
-	"overrides": [
-		{
-			"files": ["*.test.ts", "*.spec.ts"],
-			"rules": {
-				"@typescript-eslint/no-non-null-assertion": "off",
-				"@typescript-eslint/no-unsafe-return": "off",
-				"@typescript-eslint/no-unsafe-assignment": "off",
-				"@typescript-eslint/no-unsafe-member-access": "off",
-				"@typescript-eslint/no-unsafe-call": "off",
-				"@typescript-eslint/no-explicit-any": "off",
-				"max-statements": ["off"]
-			}
-		}
-	]
-}
