+++ package.json
@@ -1,83 +1,63 @@
 {
-	"name": "@lokalise/frontend-http-client",
-	"version": "2.1.0",
-	"description": "Opinionated HTTP client for the frontend",
-	"files": [
-		"dist/**",
-		"LICENSE",
-		"README.md"
-	],
-	"main": "./dist/index.cjs",
-	"types": "./dist/index.d.ts",
-	"module": "./dist/index.mjs",
-	"type": "module",
-	"exports": {
-		".": {
-			"import": {
-				"types": "./dist/index.d.ts",
-				"default": "./dist/index.mjs"
-			},
-			"require": {
-				"types": "./dist/index.d.cts",
-				"default": "./dist/index.cjs"
-			}
-		}
-	},
-	"author": {
-		"name": "Lokalise",
-		"url": "https://lokalise.com/"
-	},
-	"homepage": "https://github.com/lokalise/frontend-http-client",
-	"repository": {
-		"type": "git",
-		"url": "git://github.com/lokalise/frontend-http-client.git"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"build:dev": "tsc",
-		"build:release": "tsup",
-		"clean": "rimraf dist .eslintcache",
-		"lint": "eslint --cache --max-warnings=0 . && prettier --check src \"**/*.{json,md,ts,tsx}\" && tsc --noEmit",
-		"lint:fix": "prettier --write src \"**/*.{json,md,ts,tsx}\" --log-level=warn && eslint . --fix",
-		"test": "vitest run --coverage",
-		"prepublishOnly": "npm run clean && npm run build:release"
-	},
-	"dependencies": {
-		"fast-querystring": "^1.1.2"
-	},
-	"peerDependencies": {
-		"wretch": "^2.8.0",
-		"zod": "^3.22.0"
-	},
-	"devDependencies": {
-		"@lokalise/prettier-config": "^1.0.0",
-		"@types/node": "^20.11.5",
-		"@typescript-eslint/eslint-plugin": "^7.0.1",
-		"@typescript-eslint/parser": "^7.0.1",
-		"@vitest/coverage-v8": "^2.0.1",
-		"jest-fail-on-console": "^3.1.2",
-		"eslint": "^8.56.0",
-		"eslint-plugin-import": "^2.29.1",
-		"eslint-plugin-vitest": "^0.4.0",
-		"eslint-plugin-import-esm": "^2.0.0",
-		"mockttp": "^3.10.1",
-		"prettier": "^3.2.5",
-		"rimraf": "^5.0.5",
-		"tsup": "8.1.0",
-		"typescript": "~5.5.3",
-		"vitest": "^2.0.1"
-	},
-	"keywords": [
-		"frontend",
-		"web",
-		"browser",
-		"http",
-		"client",
-		"zod",
-		"validation",
-		"typesafe"
-	],
-	"prettier": "@lokalise/prettier-config"
+  "name": "@lokalise/frontend-http-client",
+  "version": "2.1.0",
+  "description": "Opinionated HTTP client for the frontend",
+  "files": ["dist/**", "LICENSE", "README.md"],
+  "main": "./dist/index.cjs",
+  "types": "./dist/index.d.ts",
+  "module": "./dist/index.mjs",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.mjs"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    }
+  },
+  "author": {
+    "name": "Lokalise",
+    "url": "https://lokalise.com/"
+  },
+  "homepage": "https://github.com/lokalise/frontend-http-client",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/lokalise/frontend-http-client.git"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build:dev": "tsc",
+    "build:release": "tsup",
+    "clean": "rimraf dist",
+    "lint": "biome check . && tsc --project tsconfig.lint.json --noEmit",
+    "lint:fix": "biome check --write",
+    "test": "vitest run --coverage",
+    "prepublishOnly": "npm run clean && npm run build:release"
+  },
+  "dependencies": {
+    "fast-querystring": "^1.1.2"
+  },
+  "peerDependencies": {
+    "wretch": "^2.8.0",
+    "zod": "^3.22.0"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "^1.8.3",
+    "@lokalise/biome-config": "^1.0.0",
+    "@types/node": "^20.11.5",
+    "@vitest/coverage-v8": "^2.0.1",
+    "jest-fail-on-console": "^3.1.2",
+    "mockttp": "^3.13.0",
+    "rimraf": "^5.0.8",
+    "tsup": "8.1.0",
+    "typescript": "~5.5.3",
+    "vitest": "^2.0.1"
+  },
+  "keywords": ["frontend", "web", "browser", "http", "client", "zod", "validation", "typesafe"]
 }
