+++ src/types.ts
@@ -4,113 +4,112 @@ import type { ZodSchema, z } from 'zod'
 type FreeformRecord = Record<string, unknown>
 
 export type CommonRequestParams<
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	path: string
-	responseBodySchema: ZodSchema<ResponseBody>
-	isEmptyResponseExpected?: IsEmptyResponseExpected // 204 is considered a success. Default is "false" for GET operations and "true" for everything else
-	isNonJSONResponseExpected?: IsNonJSONResponseExpected // Do not throw an error if not receiving 'application/json' content-type.  Default is "false" for GET operations and "true" for everything else
+  path: string
+  responseBodySchema: ZodSchema<ResponseBody>
+  isEmptyResponseExpected?: IsEmptyResponseExpected // 204 is considered a success. Default is "false" for GET operations and "true" for everything else
+  isNonJSONResponseExpected?: IsNonJSONResponseExpected // Do not throw an error if not receiving 'application/json' content-type.  Default is "false" for GET operations and "true" for everything else
 }
 
 export type BodyRequestParams<
-	RequestBodySchema extends z.ZodSchema,
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  RequestBodySchema extends z.ZodSchema,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	body: z.input<RequestBodySchema> | undefined
-	requestBodySchema: RequestBodySchema | undefined
+  body: z.input<RequestBodySchema> | undefined
+  requestBodySchema: RequestBodySchema | undefined
 } & CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
 
 export type FreeBodyRequestParams<
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	body?: FreeformRecord
-	requestBodySchema?: never
+  body?: FreeformRecord
+  requestBodySchema?: never
 } & CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
 
 export type QueryParams<
-	RequestQuerySchema extends z.ZodSchema,
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  RequestQuerySchema extends z.ZodSchema,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	queryParams: z.input<RequestQuerySchema> | undefined
-	queryParamsSchema: RequestQuerySchema | undefined
+  queryParams: z.input<RequestQuerySchema> | undefined
+  queryParamsSchema: RequestQuerySchema | undefined
 } & CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
 
 export type FreeQueryParams<
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	queryParams?: FreeformRecord
-	queryParamsSchema?: never
+  queryParams?: FreeformRecord
+  queryParamsSchema?: never
 } & CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
 
 export type DeleteParams<
-	RequestQuerySchema extends z.ZodSchema,
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  RequestQuerySchema extends z.ZodSchema,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	queryParams: z.input<RequestQuerySchema> | undefined
-	queryParamsSchema: RequestQuerySchema | undefined
+  queryParams: z.input<RequestQuerySchema> | undefined
+  queryParamsSchema: RequestQuerySchema | undefined
 } & Omit<
-	CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>,
-	'responseBodySchema'
+  CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>,
+  'responseBodySchema'
 > & {
-		responseBodySchema?: ZodSchema<ResponseBody>
-	}
+    responseBodySchema?: ZodSchema<ResponseBody>
+  }
 
 export type FreeDeleteParams<
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
 > = {
-	queryParams?: FreeformRecord
-	queryParamsSchema?: never
+  queryParams?: FreeformRecord
+  queryParamsSchema?: never
 } & Omit<
-	CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>,
-	'responseBodySchema'
+  CommonRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>,
+  'responseBodySchema'
 > & {
-		responseBodySchema?: ZodSchema<ResponseBody>
-	}
+    responseBodySchema?: ZodSchema<ResponseBody>
+  }
 
 export type RequestResultType<
-	ResponseBody,
-	isNonJSONResponseExpected extends boolean,
-	isEmptyResponseExpected extends boolean,
+  ResponseBody,
+  isNonJSONResponseExpected extends boolean,
+  isEmptyResponseExpected extends boolean,
 > = isEmptyResponseExpected extends true
-	? isNonJSONResponseExpected extends true
-		? WretchResponse | null
-		: ResponseBody | null
-	: isNonJSONResponseExpected extends true
-		? WretchResponse
-		: ResponseBody
+  ? isNonJSONResponseExpected extends true
+    ? WretchResponse | null
+    : ResponseBody | null
+  : isNonJSONResponseExpected extends true
+    ? WretchResponse
+    : ResponseBody
 
 export type ResourceChangeParams<
-	RequestBody,
-	ResponseBody,
-	IsNonJSONResponseExpected extends boolean,
-	IsEmptyResponseExpected extends boolean,
-	RequestQuerySchema extends z.Schema | undefined = undefined,
+  RequestBody,
+  ResponseBody,
+  IsNonJSONResponseExpected extends boolean,
+  IsEmptyResponseExpected extends boolean,
+  RequestQuerySchema extends z.Schema | undefined = undefined,
 > = (RequestBody extends z.Schema
-	? BodyRequestParams<RequestBody, ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
-	: FreeBodyRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>) &
-	(RequestQuerySchema extends z.Schema
-		? QueryParams<
-				RequestQuerySchema,
-				ResponseBody,
-				IsNonJSONResponseExpected,
-				IsEmptyResponseExpected
-			>
-		: FreeQueryParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>)
+  ? BodyRequestParams<RequestBody, ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>
+  : FreeBodyRequestParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>) &
+  (RequestQuerySchema extends z.Schema
+    ? QueryParams<
+        RequestQuerySchema,
+        ResponseBody,
+        IsNonJSONResponseExpected,
+        IsEmptyResponseExpected
+      >
+    : FreeQueryParams<ResponseBody, IsNonJSONResponseExpected, IsEmptyResponseExpected>)
 
-// We don't know which addons Wretch will have, and we don't really care, hence any
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
+// biome-ignore lint/suspicious/noExplicitAny: We don't know which addons Wretch will have, and we don't really care, hence any
 export type WretchInstance = Wretch<any, unknown, undefined>
