+++ src/utils/either.ts
@@ -1,11 +1,11 @@
 type Left<T> = {
-	error: T
-	result?: never
+  error: T
+  result?: never
 }
 
 type Right<U> = {
-	error?: never
-	result: U
+  error?: never
+  result: U
 }
 
 /**
@@ -21,16 +21,16 @@ export type Either<T, U> = NonNullable<Left<T> | Right<U>>
  * Variation of Either, which may or may not have Error set, but always has Result
  */
 export type DefiniteEither<T, U> = {
-	error?: T
-	result: U
+  error?: T
+  result: U
 }
 
 export const isFailure = <T, U>(e: Either<T, U>): e is Left<T> => {
-	return e.error !== undefined
+  return e.error !== undefined
 }
 
 export const isSuccess = <T, U>(e: Either<T, U>): e is Right<U> => {
-	return e.result !== undefined
+  return e.result !== undefined
 }
 
 export const failure = <T>(error: T): Left<T> => ({ error })
