+++ tsconfig.json
@@ -1,26 +1,26 @@
 {
-	"compilerOptions": {
-		"outDir": "dist",
-		"module": "node16",
-		"target": "ES2022",
-		"lib": ["ES2022", "dom"],
-		"sourceMap": false,
-		"declaration": true,
-		"declarationMap": false,
-		"types": ["node", "vitest/globals"],
-		"strict": true,
-		"moduleResolution": "node16",
-		"noUnusedLocals": false,
-		"noUnusedParameters": false,
-		"noFallthroughCasesInSwitch": true,
-		"strictNullChecks": true,
-		"importHelpers": true,
-		"baseUrl": ".",
-		"allowSyntheticDefaultImports": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"resolveJsonModule": true
-	},
-	"include": ["src/**/*", "index.ts"],
-	"exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
+  "compilerOptions": {
+    "outDir": "dist",
+    "module": "node16",
+    "target": "ES2022",
+    "lib": ["ES2022", "dom"],
+    "sourceMap": false,
+    "declaration": true,
+    "declarationMap": false,
+    "types": ["node", "vitest/globals"],
+    "strict": true,
+    "moduleResolution": "node16",
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "noFallthroughCasesInSwitch": true,
+    "strictNullChecks": true,
+    "importHelpers": true,
+    "baseUrl": ".",
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true
+  },
+  "include": ["src/**/*", "index.ts"],
+  "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
 }
