+++ tsconfig.lint.json
@@ -1,5 +1,5 @@
 {
-	"extends": ["./tsconfig.json"],
-	"include": ["./src/**/*", "index.ts", "vitest.config.ts", "tsup.config.ts"],
-	"exclude": []
+  "extends": ["./tsconfig.json"],
+  "include": ["./src/**/*", "index.ts", "vitest.config.ts", "tsup.config.ts"],
+  "exclude": []
 }
